Relocate header content into header file
